refactor: remove unnecessary .toInt() on int Offset coordinates and use List.filled for constant cell initialization
Offset.x and Offset.y are already int in RadarTUI (terminal cells are discrete),
so .toInt() was redundant noise. Also replaced List.generate with List.filled
in OutputBuffer.resize() since all cells are initialized to the same constant value.

+++ lib/src/widgets/basic/card.dart
@@ -170,8 +170,7 @@ class RenderCard extends RenderBox with RenderObjectWithChildMixin<RenderBox> {
     if (color == null) return;
     final TextStyle bgStyle = _cachedBorderStyle!;
     for (int y = 0; y < height; y++) {
-      context.fillBackground(
-          offset.x.toInt(), offset.y.toInt() + y, width, bgStyle);
+      context.fillBackground(offset.x, offset.y + y, width, bgStyle);
     }
   }
 
@@ -229,7 +228,7 @@ class RenderCard extends RenderBox with RenderObjectWithChildMixin<RenderBox> {
     int height,
     TextStyle style,
   ) {
-    final int y = offset.y.toInt() + height - 1;
+    final int y = offset.y + height - 1;
     context.writeString(
       offset.x,
       y,
